function get_radio_field_title($doc_id, $field_name) {     // Define the query arguments     $args = array(         'post_type'      => 'doc-field',         'posts_per_page' => 1, // We only need one post since we're looking for a specific field         'meta_query'     => array(             'relation' => 'AND',             array(                 'key'   => 'doc_id',                 'value' => $doc_id,             ),             array(                 'key'   => 'field_name',                 'value' => $field_name,             ),         ),     );
    // Perform the query
    $query = new WP_Query($args);

    // Check if there are any posts found
    if ($query->have_posts()) {
        // Retrieve the post
        $query->the_post();

        // Get the value of the 'field_title' meta key
        $field_title = get_post_meta(get_the_ID(), 'field_title', true);

        // Reset post data
        wp_reset_postdata();

        // Return the field title
        return $field_title;
    } else {
        // If no matching post is found, return false or any default value
        return false;
    }
}

+++ includes/display-profiles.php
@@ -680,7 +680,7 @@ add_action( 'wp_ajax_nopriv_get_site_list_data', 'get_site_list_data' );
 function get_site_dialog_data() {
     $response = array();
     if( isset($_POST['_site_id']) ) {
-        $site_id = (int)sanitize_text_field($_POST['_site_id']);
+        $site_id = sanitize_text_field($_POST['_site_id']);
         $response["site_title"] = get_the_title($site_id);
     }
     wp_send_json($response);
